src/br/com/thborja/service/ContratoService.java
package br.com.thborja.service;

import br.com.thborja.dao.IContratoDao;

public class ContratoService implements IContratoService {

    private IContratoDao contratoDao;

    public ContratoService(IContratoDao mock2) {
        this.contratoDao = mock2;
    }

    @Override
    public String salvar() {
        //throw new UnsupportedOperationException();
        contratoDao.salvar();
        return "Sucesso";
    }

    @Override
    public String buscar() {
        contratoDao.buscar();
        return "Sucesso";
    }

    @Override
    public String excluir() {
        contratoDao.excluir();
        return "Sucesso";
    }

    @Override
    public String atualizar() {
        contratoDao.atualizar();
        return "Sucesso";
    }
}
